requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
dnsdumpster==0.10
dnspython==2.6.1
idna==3.7
kombu==5.3.7
prompt_toolkit==3.0.46
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
